# proyecto3/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Banco XYZ Services
  description: API para gestionar transacciones del Banco XYZ
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local
paths:
  /clientes:
    get:
      tags:
        - cliente
      summary: Listar clientes
      operationId: listCustomers.
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
        default:
          description: Error en la obtención de clientes
    post:
      tags:
        - cliente
      summary: Crear un cliente
      operationId: registerCustomer
      requestBody:
        description: Registrar un cliente
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
      responses:
        '200':
          description: Cliente registrado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        default:
          description: Error en la creacion de cliente
  /clientes/{id}:
    get:
      tags:
        - cliente
      summary: Obtener datos de un cliente
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cliente obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        default:
          description: Error en obtener datos del cliente
    put:
      tags:
        - cliente
      summary: Actualizar un cliente
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Detalles del cliente actualizada
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
      responses:
        '200':
          description: Cliente actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        default:
          description: Error en la actualización de cliente
    delete:
      tags:
        - cliente
      summary: Eliminar cliente por su ID
      operationId: deleteCustomerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Cliente eliminado
        default:
          description: Error al eliminar el cliente
...
components:
  schemas:
    CustomerRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único del cliente
        nombre:
          type: string
          description: Nombre del cliente
        apellido:
          type: string
          description: Apellido del cliente
        dni:
          type: string
          description: DNI del cliente
        email:
          type: string
          description: Email del cliente
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
          description: Nombre del cliente
        apellido:
          type: string
          description: Apellido del cliente
        dni:
          type: string
          description: DNI del cliente
        email:
          type: string
          description: Email del cliente
...